API Documentation

Example

Let's consider the following TimeSpan

julia> using TimeSpans, AlignedSpans, DatesERROR: ArgumentError: Package TimeSpans not found in current path:
- Run `import Pkg; Pkg.add("TimeSpans")` to install the TimeSpans package.
julia> span = TimeSpan(Millisecond(1500), Millisecond(3500))ERROR: UndefVarError: Millisecond not defined

If we have a 1 Hz signal, there are various ways we can index into it using this TimeSpan. One option is to round the endpoints down:

julia> down_span = AlignedSpan(1, span, RoundEndsDown)ERROR: UndefVarError: AlignedSpan not defined
julia> n_samples(down_span)ERROR: UndefVarError: n_samples not defined

The second sample of our signal occurs at time 1s (since we have a 1Hz signal that starts at 0s). When we round the starting endpoint down from 1.5s to the nearest sample, we find that sample. This can be seen as "the last sample that occurred before time 1.5s".

Perhaps instead we would like to round the endpoints "inward" to only consider samples occurring with the time span:

julia> in_span = AlignedSpan(1, span, RoundInward)ERROR: UndefVarError: AlignedSpan not defined
julia> n_samples(in_span)ERROR: UndefVarError: n_samples not defined

Interface

AlignedSpans.AlignedSpan — Type

AlignedSpan(sample_rate, span, mode::EndpointRoundingMode)

Creates an AlignedSpan by rounding the left endpoint according to mode.start, and the right endpoint by mode.stop.

If mode.start==RoundUp, then the left index of the resulting span is guaranteed to be inside span. This is accomplished by checking if the left endpoint of the span is exclusive, and if so, incrementing the index after rounding when necessary.

Likewise, if mode.start==RoundDown, then the right index of the resulting span is guaranteed to be inside span. This is accomplished by checking if the right endpoint of the span is exclusive, and if so, decrementing the index after rounding when necessary.

Note: span may be of any type which which provides methods for AlignedSpans.start_index_from_time and AlignedSpans.stop_index_from_time.

AlignedSpan(sample_rate, span, mode::ConstantSamplesRoundingMode)

Creates an AlignedSpan whose left endpoint is rounded according to mode.start, and whose right endpoint is determined so by the left endpoint and the total number of samples, given by AlignedSpans.n_samples(sample_rate, duration(span)).

Note: span may be of any type which which provides a method for AlignedSpans.start_index_from_time and AlignedSpans.duration.

AlignedSpans.start_index_from_time — Function

start_index_from_time(sample_rate, span, rounding_mode)

Returns the index of a sample object obtained by rounding the start of span according to rounding_mode.

AlignedSpans.stop_index_from_time — Function

stop_index_from_time(sample_rate, span, rounding_mode)

Returns the index of a sample object obtained by rounding the stop of span according to rounding_mode.

AlignedSpans.duration — Function

duration(span)

Return the duration of span.

AlignedSpans.EndpointRoundingMode — Type

EndpointRoundingMode(start::RoundingMode, stop::RoundingMode)

Creates a rounding object for AlignedSpan to indicate how the AlignedSpan's endpoints should be determined from a given spans endpoints'.

AlignedSpans.ConstantSamplesRoundingMode — Type

ConstantSamplesRoundingMode(start::RoundingMode)

Creates a rounding object for AlignedSpan to indicate the AlignedSpan should be constructed by the start and duration of the span, without regard to its stop.

If two spans have the same duration, then the resulting AlignedSpan's will have the same number of samples when constructed with this rounding mode.

Utilities

AlignedSpans.n_samples — Function

n_samples(sample_rate, duration::Period)

Returns the minimal number of samples that can occur in a span of duration duration.

n_samples(aligned::AlignedSpan)

Returns the number of samples present in the span aligned.

AlignedSpans.consecutive_subspans — Function

consecutive_subspans(span::AlignedSpan, duration::Period)

Creates an iterator of AlignedSpan such that each AlignedSpan has consecutive indices which cover all of the original span's indices. In particular,

Each span has n = n_samples(span.sample_rate, duration) samples, except possibly

the last one, which may have fewer.

The number of subspans is given by cld(n_samples(span), n)

The number of samples in the last subspan is r = rem(n_samples(span), n) unless r=0, in which case the the last subspan has the same number of samples as the rest, namely n.