API documentation

AlignedSpans.EndpointRoundingMode — Type

EndpointRoundingMode(start::RoundingMode, stop::RoundingMode)

Creates a rounding object for AlignedSpan to indicate how the AlignedSpan's endpoints should be determined from a given spans endpoints'.

AlignedSpans.AlignedSpan — Method

AlignedSpan(sample_rate, span, mode::EndpointRoundingMode)

Creates an AlignedSpan by rounding the left endpoint according to mode.start, and the right endpoint by mode.stop.

If mode.start==RoundUp, then the left index of the resulting span is guaranteed to be inside span. This is accomplished by checking if the left endpoint of the span is exclusive, and if so, incrementing the index after rounding when necessary.

Likewise, if mode.start==RoundDown, then the right index of the resulting span is guaranteed to be inside span. This is accomplished by checking if the right endpoint of the span is exclusive, and if so, decrementing the index after rounding when necessary.

Note: span may be of any type which which provides methods for AlignedSpans.start_index_from_time and AlignedSpans.stop_index_from_time.

AlignedSpans.ConstantSamplesRoundingMode — Type

ConstantSamplesRoundingMode(start::RoundingMode)

Creates a rounding object for AlignedSpan to indicate the AlignedSpan should be constructed by the start and duration of the span, without regard to its stop.

If two spans have the same duration, then the resulting AlignedSpan's will have the same number of samples when constructed with this rounding mode.

See also AlignedSpan(sample_rate, span, mode::ConstantSamplesRoundingMode).

AlignedSpans.AlignedSpan — Method

AlignedSpan(sample_rate, span, mode::ConstantSamplesRoundingMode)

Creates an AlignedSpan whose left endpoint is rounded according to mode.start, and whose right endpoint is determined so by the left endpoint and the number of samples, given by AlignedSpans.n_samples(sample_rate, duration(span)).

Interface: span may be of any type which which provides a method for AlignedSpans.start_index_from_time and TimeSpans.duration.

More detailed information

This is designed so that if AlignedSpan(sample_rate, span, mode::ConstantSamplesRoundingMode) is applied to multiple spans, with the same sample_rate, and the same durations, then the resulting AlignedSpan's will have the same number of samples.

For this reason, we ask for TimeSpans.duration(span) to be defined, rather than a n_samples(span) function: the idea is that we want to only using the duration and the starting time, rather than the actual number of samples in this particular span.

In contrast, AlignedSpan(sample_rate, span, RoundInward) provides an AlignedSpan which includes only (and exactly) the samples contained within span.

If one wants to create a collection of consecutive, non-overlapping, AlignedSpans each with the same number of samples, then use consecutive_subspans instead.

AlignedSpans.consecutive_subspans — Function

consecutive_subspans(span::AlignedSpan, duration::Period)

Creates an iterator of AlignedSpan such that each AlignedSpan has consecutive indices which cover all of the original span's indices. In particular,

Each span has n = n_samples(span.sample_rate, duration) samples, except possibly

the last one, which may have fewer.

The number of subspans is given by cld(n_samples(span), n)

The number of samples in the last subspan is r = rem(n_samples(span), n) unless r=0, in which case the the last subspan has the same number of samples as the rest, namely n.

AlignedSpans.n_samples — Function

n_samples(sample_rate, duration::Period)

Returns the minimal number of samples that can occur in a span of duration duration.

n_samples(aligned::AlignedSpan)

Returns the number of samples present in the span aligned.

AlignedSpans.indices — Function

AlignedSpans.indices(span::AlignedSpan) -> UnitRange{Int64}

Returns the sample indices associated to an AlignedSpan.

Interface for conversion from continuous time spans

In order to support conversion of continuous time span types to AlignedSpan's, three methods may be defined. These are not exported, because they are generally not used directly, but rather defined in order to facilitate use of the AlignedSpan constructors.

AlignedSpans.start_index_from_time — Function

AlignedSpans.start_index_from_time(sample_rate, span, rounding_mode)

Returns the index of a sample object obtained by rounding the start of span according to rounding_mode.

See also AlignedSpan(sample_rate, span, mode::EndpointRoundingMode) and AlignedSpan(sample_rate, span, mode::ConstantSamplesRoundingMode).

AlignedSpans.stop_index_from_time — Function

AlignedSpans.stop_index_from_time(sample_rate, span, rounding_mode)

Returns the index of a sample object obtained by rounding the stop of span according to rounding_mode.

See also AlignedSpan(sample_rate, span, mode::EndpointRoundingMode).

Missing docstring.

Missing docstring for TimeSpans.duration. Check Documenter's build log for details.